Weaken int casts (#197)
* Change `as_s32()` and `as_u32()` to be silent

Diffs from this change in OOT, MM, PM:

https://gist.github.com/zbanks/d8655acae883aa80a01c265c7acde0f8

* Use weaker (unsigned) int casts instead of s32/u32

Diffs for PM, OOT, and MM here:

https://gist.github.com/zbanks/674e4f51171aa3913a36e4dc2f981034

* Unify left/right types in `Binary.[isu]cmp(...)`

Diffs from OOT, MM, PM here (large):

https://gist.github.com/zbanks/f3ba05509f4e83e429bf4e137aabfe9e

* Revert "Unify left/right types in `Binary.[isu]cmp(...)`"

This reverts commit 447423bf2678f6bfba0ddc0cb07b5d682e7ef268.

* Revert changes to silent for as_s32/as_u32

* Replace most s32/u32 with sintish/uintish

* Remove unused as_s32; fix silent

+++ tests/end_to_end/gcc-division/mwcc-o4p-out.c
@@ -20,7 +20,7 @@ void test_s8(s8 c) {
     s8 temp_r4_3;
 
     sp8 = c;
-    func_00400090((u32) ((s32) (s8) (u8) sp8 / 2));
+    func_00400090((u32) ((s8) (u8) sp8 / 2));
     func_00400090(MULT_HI(0x55555556, (s8) (u8) sp8) + ((s8) (u8) sp8 / 6442450941));
     func_00400090((s8) (u8) sp8 / 5);
     temp_r0 = (s8) (u8) sp8;
@@ -57,7 +57,7 @@ void test_s16(s16 h) {
     s32 temp_r3;
 
     sp8 = h;
-    func_00400090((u32) ((s32) sp8 / 2));
+    func_00400090((u32) (sp8 / 2));
     func_00400090(MULT_HI(0x55555556, sp8) + (sp8 / 6442450941));
     func_00400090(sp8 / 5);
     temp_r0 = (s32) (MULT_HI(0x92492493, sp8) + sp8) >> 2;
